backend/requirements.txt
# AI-Guardian - Backend Python Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
motor==3.5.1
pymongo==4.9.1
paho-mqtt==1.6.1
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
websockets==12.0
numpy==1.26.4
scikit-learn==1.5.2
